# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DraftStartBody": {
        "properties": {
          "forever": {
            "default": false,
            "title": "Forever",
            "type": "boolean"
          },
          "interval_s": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 0.5,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval S"
          },
          "season": {
            "default": 2026,
            "title": "Season",
            "type": "integer"
          }
        },
        "title": "DraftStartBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegenBody": {
        "properties": {
          "baseline": {
            "default": "live",
            "pattern": "^(live|historical)$",
            "title": "Baseline",
            "type": "string"
          },
          "provider": {
            "default": "ensemble",
            "pattern": "^(sleeper|espn|ensemble)$",
            "title": "Provider",
            "type": "string"
          },
          "season": {
            "default": 2026,
            "title": "Season",
            "type": "integer"
          }
        },
        "title": "RegenBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Lazy Sleeper API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/board": {
      "get": {
        "description": "The latest persisted board: ranked rows with points, VORP, tier/cliff, ADP delta,\ndisagreement spread/flag and injury status. `position` filters; `rank` stays overall.",
        "operationId": "board_board_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "default": 2026,
              "title": "Season",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "default": "ensemble",
              "pattern": "^(sleeper|espn|ensemble)$",
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "position",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 8,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Position"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Board Board Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Board"
      }
    },
    "/board.html": {
      "get": {
        "description": "Self-contained HTML view of the latest board \u2014 the draft-night fallback.",
        "operationId": "board_html_board_html_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "default": 2026,
              "title": "Season",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "default": "ensemble",
              "pattern": "^(sleeper|espn|ensemble)$",
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Board Html"
      }
    },
    "/board/config": {
      "get": {
        "description": "Tier/cliff/flag thresholds in force (defaults seeded by migrations 0006/0007).",
        "operationId": "board_config_board_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Board Config Board Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Board Config"
      },
      "put": {
        "description": "Adjust any subset of the tier/cliff/flag thresholds (draft-day dial).",
        "operationId": "put_board_config_board_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardConfigBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Board Config Board Config Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Board Config"
      }
    },
    "/board/config.html": {
      "get": {
        "description": "Tuning page: every `board_config` dial as a form \u2192 `PUT /board/config`, then apply to\nthe running draft (`POST /draft/{id}/config`) \u2014 signal dials instantly, board-time dials\nvia a restart. Linked from the draft page so no CLI is needed on draft night.",
        "operationId": "board_config_html_board_config_html_get",
        "parameters": [
          {
            "in": "query",
            "name": "draft_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Board Config Html"
      }
    },
    "/board/regen": {
      "post": {
        "description": "Rebuild the board from core.* under the current config and persist it \u2014 on-demand\nregen, so a `PUT /board/config` change shows up without waiting for the daily job.",
        "operationId": "board_regen_board_regen_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Board Regen Board Regen Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Board Regen"
      }
    },
...
    "/draft.html": {
      "get": {
        "description": "`/draft/{id}/state.html` for the configured `sleeper_draft_id` \u2014 the bookmark.",
        "operationId": "draft_html_default_draft_html_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "default": 2026,
              "title": "Season",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 40,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": 2.0,
              "maximum": 60.0,
              "minimum": 0.5,
              "title": "Refresh",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": 2.0,
              "maximum": 60.0,
              "minimum": 0.5,
              "title": "Interval",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Html Default"
      }
    },
...
    "/draft/{draft_id}/start": {
      "post": {
        "description": "Pre-draft load (board + ADP + player lookup, once) and start polling Sleeper for this\ndraft on a background thread. Idempotent while the runner is alive. Do this *before* the\ndraft room opens \u2014 the board build takes seconds, the recompute per pick ~50 ms.",
        "operationId": "draft_start_draft__draft_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DraftStartBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftStartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Start"
      }
    },
...
  }
}